pyproject.toml
[tool.poetry]
name = "queuery-client"
version = "0.1.0"
description = "Queuery Redshift HTTP API Client for Python"
authors = ["altescy <altescy@fastmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.1"
requests = "^2.25.1"
importlib-metadata = "^3.7.2"
python-dateutil = "^2.8.2"
pandas = {version = "^1.2.3", optional = true}

[tool.poetry.dev-dependencies]
python-language-server = "^0.36.2"
pysen = {version = "^0.10.1", extras = ["lint"]}
pytest = "^6.2.4"

[tool.poetry.extras]
pandas = ["pandas"]

[tool.pysen]
version = "0.10"

[tool.pysen-cli]
settings_dir = "."

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 115
py_version = "py38"
[[tool.pysen.lint.mypy_targets]]
  paths = ["."]

[tool.pysen.lint.source]
  includes = ["."]
  excludes = [".venv/"]

[tool.black] # automatically generated by pysen
line-length = 115
target-version = ["py38"]

[tool.isort] # automatically generated by pysen
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
line_length = 115
multi_line_output = 3
use_parentheses = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
